+++ zerver/tests/test_events.py
@@ -1841,7 +1841,7 @@ class NormalActionsTest(BaseAction):
                 backend, "can_mention_groups", moderators_group, acting_user=None
             )
         )
-        check_user_group_update("events[0]", events[0], "can_mention_group")
+        check_user_group_update("events[0]", events[0], "can_mention_groups")
 
         # Test add members
         hamlet = self.example_user("hamlet")